requirements.txt
anthropic>=0.39.0
streamlit>=1.40.0
pandas>=2.0.0
python-docx>=1.1.0
plotly>=5.18.0
